docs: Update README
More detailed README to explain and summarize the description of this project in more detail that was worked on in General Assembly. We were to analyze a database using SQL queries in order to make recommendations on how to improve liquor sales. The SQL questions and queries are contained in the text document, the charts used for the presentations are located in the Excel file, and my presentation to this project is in the PowerPoint file.